feat(SEC9): OpenShell is the only runtime — remove microsandbox + in-process production path
OpenShell has soaked as the default; this removes the alternatives.

Plugins: delete microsandbox-runtime.ts, microsandbox-sdk.ts, and the
`microsandbox` npm dep. The runtime-agnostic abstractions (PluginRuntime,
PluginVmSpec, networkModeFor, buildExecCommand) move to a neutral
plugin-runtime.ts that OpenShellRuntime and the dev SubprocessRuntime
share. createDefaultRuntime always returns OpenShellRuntime; the
OPENNEKO_PLUGIN_RUNTIME escape hatch is gone.

Agent: the OPENNEKO_AGENT_RUNTIME gate is removed from the broker
(always started), the sandbox launcher deps (always injected), host
provisioning, and the web run route. Tests still exercise the in-process
core via injected deps.runCore — that path stays for tests only.

CLI/compose: drop the `--runtime` flag, the inprocess.yml and
plugins.linux.yml + compose.plugins.yml overlays, and the AgentRuntime/
HasKVM supervisor fields. openshell.yml always applies. The host check
now requires Docker (containers) instead of KVM (microsandbox-era), so
amd64 macs are supported again; doctor/install messaging updated.

Docs: README/INSTALL/CONTRIBUTING reflect sandbox-by-default; no
libkrun/KVM/--runtime references remain in user-facing files.

Tests green: llm 516, worker 260, web 232, Go all packages. SEC9 done.
