+++ README.md
@@ -86,8 +86,8 @@ openneko install @open-neko/plugin-parallel-search
 
 - **Self-hosted via Docker.** One binary, one `start` command. Data lives on your Postgres.
 - **Bring your own LLM.** Hermes runs against Anthropic, OpenAI, Google, Ollama, and others; Claude Agent runs Anthropic in-process. Pick per task, swap any time.
-- **Plugins in microVM sandboxes.** Outbound network is allowlisted per manifest, not blanket-open.
-- **Agent-in-sandbox (preview).** `openneko start --runtime openshell` runs the agent loop itself inside a policy sandbox — default-deny egress, and the model API key never enters the box (the gateway proxy injects it on the wire). See [OPENSHELL.md](OPENSHELL.md).
+- **Plugins in OpenShell sandboxes.** Outbound network is allowlisted per manifest, not blanket-open.
+- **Agent in a sandbox by default.** The agent loop itself runs inside an OpenShell policy sandbox — default-deny egress, and the model API key never enters the box (the gateway proxy injects it on the wire). See [OPENSHELL.md](OPENSHELL.md).
 - **Apache 2.0.** Inspect everything. Fork it. Take your data with you.
 
 ## Docs
