--- apps/openneko/assets/compose/inprocess.yml
@@ -1,12 +1,0 @@
-# Dev escape hatch (`openneko start --runtime inprocess`). The code
-# defaults to OpenShell (SEC11), so opting out has to say so explicitly —
-# otherwise worker/web would look for a gateway that isn't in the stack.
-# Removed with SEC9 when the in-process production path goes away.
-services:
-  worker:
-    environment:
-      OPENNEKO_AGENT_RUNTIME: inprocess
-      OPENNEKO_PLUGIN_RUNTIME: microsandbox
-  web:
-    environment:
-      OPENNEKO_AGENT_RUNTIME: inprocess
