+++ apps/openneko/internal/host/host.go
@@ -1,9 +1,7 @@
 package host
 
 import (
-	"errors"
-	"io/fs"
-	"os"
+	"os/exec"
 	"runtime"
 )
 
@@ -13,52 +11,42 @@ type Result struct {
 	Reason    string
 }
 
-// Check reports whether microsandbox is expected to run on this host. Mirrors
-// the gating in apps/openneko-cli/src/host-check.ts and worker/src/plugins/
-// microsandbox-sdk.ts; keep in sync if microsandbox adds targets.
+// Check reports whether the OpenShell sandbox runtime is expected to run on
+// this host (SEC9: OpenShell is the only runtime). The gateway and every
+// sandbox run as containers, so the requirement is Docker — not KVM, which
+// was a microsandbox-era constraint.
 func Check() Result {
-	return checkWith(runtime.GOOS, runtime.GOARCH, hasKVM)
+	return checkWith(runtime.GOOS, runtime.GOARCH, hasDocker)
 }
 
-func checkWith(goos, goarch string, kvm func() bool) Result {
+func checkWith(goos, goarch string, docker func() bool) Result {
+	triple := goos + "-" + goarch
 	switch goos {
-	case "darwin":
-		if goarch != "arm64" {
-			return Result{
-				Supported: false,
-				Triple:    goos + "-" + goarch,
-				Reason:    "macOS x86_64 is not supported — microsandbox bundles only arm64. Run OpenNeko on Apple Silicon, or on a Linux host with KVM.",
-			}
-		}
-		return Result{Supported: true, Triple: "darwin-arm64"}
-	case "linux":
-		triple := "linux-x64-gnu"
-		if goarch == "arm64" {
-			triple = "linux-arm64-gnu"
-		} else if goarch != "amd64" {
-			return Result{
-				Supported: false,
-				Triple:    goos + "-" + goarch,
-				Reason:    "Linux " + goarch + " not supported by the bundled microsandbox runtime.",
-			}
-		}
-		if !kvm() {
+	case "darwin", "linux":
+		if goarch != "arm64" && goarch != "amd64" {
 			return Result{
 				Supported: false,
 				Triple:    triple,
-				Reason:    "/dev/kvm not found — microsandbox needs KVM (nested virtualization in cloud VMs may need the host enabled).",
+				Reason:    goos + " " + goarch + " is not supported — the OpenShell sandbox images ship for amd64 and arm64.",
+			}
+		}
+		if !docker() {
+			return Result{
+				Supported: false,
+				Triple:    triple,
+				Reason:    "docker not found on PATH — the OpenShell gateway and plugin sandboxes run as containers.",
 			}
 		}
 		return Result{Supported: true, Triple: triple}
 	}
 	return Result{
 		Supported: false,
-		Triple:    goos + "-" + goarch,
+		Triple:    triple,
 		Reason:    "Windows is not currently supported. WSL2 viability is being evaluated.",
 	}
 }
 
-func hasKVM() bool {
-	_, err := os.Stat("/dev/kvm")
-	return !errors.Is(err, fs.ErrNotExist)
+func hasDocker() bool {
+	_, err := exec.LookPath("docker")
+	return err == nil
 }
