+++ apps/web/vitest.config.ts
@@ -16,8 +16,6 @@ export default defineConfig({
     // Production defaults to openshell (SEC11); tests exercise the
     // in-process core. Openshell-path tests override per-test.
     env: {
-      OPENNEKO_AGENT_RUNTIME: "inprocess",
-      OPENNEKO_PLUGIN_RUNTIME: "microsandbox",
     },
     alias: {
       // Match the Next.js TS path alias from tsconfig.json.
