+++ apps/worker/vitest.config.ts
@@ -15,8 +15,6 @@ export default defineConfig({
     // Production defaults to openshell (SEC11); tests exercise the
     // in-process core. Openshell-path tests override per-test.
     env: {
-      OPENNEKO_AGENT_RUNTIME: "inprocess",
-      OPENNEKO_PLUGIN_RUNTIME: "microsandbox",
     },
   },
 });
